[PostGIS] Reviewed big patch from Bruno Scott including fixes for: [https://trac.osgeo.org/fdo/ticket/94], [https://trac.osgeo.org/fdo/ticket/106], [https://trac.osgeo.org/fdo/ticket/171], [https://trac.osgeo.org/fdo/ticket/178], [https://trac.osgeo.org/fdo/ticket/232], [https://trac.osgeo.org/fdo/ticket/233], [https://trac.osgeo.org/fdo/ticket/234], [https://trac.osgeo.org/fdo/ticket/235], [https://trac.osgeo.org/fdo/ticket/236], [https://trac.osgeo.org/fdo/ticket/241], [https://trac.osgeo.org/fdo/ticket/310], [https://trac.osgeo.org/fdo/ticket/311], [https://trac.osgeo.org/fdo/ticket/312], [https://trac.osgeo.org/fdo/ticket/313], [https://trac.osgeo.org/fdo/ticket/314]. Thanks to Bruno and his Team for the fantastic work.
git-svn-id: http://svn.osgeo.org/fdo/trunk@3904 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/PostGIS/Src/Provider/ExpressionProcessor.cpp
@@ -280,17 +280,17 @@ void ExpressionProcessor::ProcessDateTimeValue(FdoDateTimeValue& expr)
         
         if (dt.IsTime())
         {
-            value = str(boost::format("'%d:%d:%d'") % dt.hour % dt.minute % dt.seconds);
+            value = str(boost::format("'%d:%d:%d'") % (int)dt.hour % (int)dt.minute % (int)dt.seconds);
             format = "'HH24:MI:SS'";
         }
         else if (dt.IsDate())
         {
-            value = str(boost::format("'%d-%d-%d'") % dt.month % dt.day % dt.year);
+            value = str(boost::format("'%d-%d-%d'") % (int)dt.month % (int)dt.day % (int)dt.year);
             format = "'MM-DD-YYYY'";
         }
         else if (dt.IsDateTime())
         {
-            value = str(boost::format("'%d-%d-%d %d:%d:%d'") % dt.month % dt.day % dt.year % dt.month % dt.day % dt.year);
+            value = str(boost::format("'%d-%d-%d %d:%d:%d'") % (int)dt.month % (int)dt.day % (int)dt.year % (int)dt.month % (int)dt.day % (int)dt.year);
             format = "'MM-DD-YYYY HH24:MI:SS'";
         }
         else
